# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTFUL Laravel 6.0",
        "description": "API documentation project",
        "contact": {
            "email": "carlos.clayton@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/password": {
            "post": {
                "tags": [
                    "Autentication"
                ],
                "summary": "Change password",
                "description": "Receive an email to change password",
                "operationId": "App\\Http\\Controllers\\Auth\\ForgotPasswordController::password",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/password/forgot": {
            "post": {
                "tags": [
                    "Autentication"
                ],
                "summary": "Forgot password",
                "description": "Receive an email to change password",
                "operationId": "App\\Http\\Controllers\\Auth\\ForgotPasswordController::sendResetLinkEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/access_token": {
            "post": {
                "tags": [
                    "Autentication"
                ],
                "summary": "Get token",
                "description": "Return token",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::accessToken",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autentication"
                ],
                "summary": "Logout ",
                "description": "Logout",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/refresh_token": {
            "post": {
                "tags": [
                    "Autentication"
                ],
                "summary": "Refresh token",
                "description": "Refresh token",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::refreshToken",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autentication"
                ],
                "summary": "Register new user",
                "description": "Register new user",
                "operationId": "App\\Http\\Controllers\\Auth\\RegisterController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List of categories",
                "description": "Return a list of categories",
                "operationId": "App\\Http\\Controllers\\CategoriesController::index",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "summary": "Store a category",
                "description": "Return message",
                "operationId": "App\\Http\\Controllers\\CategoriesController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store categories"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "get": {
                "summary": "Show a category",
                "description": "Return a category",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "summary": "Update a category",
                "description": "Update a category",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store categories"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "scheme": "http"
            }
        }
    }
}